fixes import file, completes merchant BI

+++ db/migrate/20150915030904_add_column_unit_price_to_items.rb
@@ -1,5 +1,5 @@
 class AddColumnUnitPriceToItems < ActiveRecord::Migration
   def change
-    add_column :items, :unit_price, :float
+    add_column :items, :unit_price, :decimal
   end
 end
