+++ db/migrate/20150915031023_add_column_unit_price_to_invoice_items.rb
@@ -1,5 +1,5 @@
 class AddColumnUnitPriceToInvoiceItems < ActiveRecord::Migration
   def change
-    add_column :invoice_items, :unit_price, :float
+    add_column :invoice_items, :unit_price, :decimal
   end
 end
